requirements.txt
asgiref==3.5.0
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
idna==3.3
numpy==1.22.3
psycopg2==2.9.3
pycodestyle==2.8.0
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
style==1.1.0
toml==0.10.2
tzdata==2021.5
update==0.0.1
urllib3==1.26.8
